ALL in python script
This code takes in a search term related to certain disease. It crawls PubMed for relevant new publications (the number can be customized). Then it downloads them and using AI tensor flow looks for the answer to the question the user posed. Then finally display the answers on the screen. The coming update will include a user-friendly interface wrapped in docker to eliminate problems. If you use this program for drug repurposing or something else please cite 
Norwin Kubick et al. Molecules. 2020 May; 25(9): 2155.

Please feel free to try it. In case you notice nay bugs please email us so we can check it out for you!